+++ debug.log
@@ -4,3 +4,5 @@
 [0222/011358:ERROR:node_debugger.cc(86)] Cannot start debugger server
 [0222/093238:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:6004
 [0222/093238:ERROR:node_debugger.cc(86)] Cannot start debugger server
+[0222/171151:ERROR:tcp_listen_socket.cc(76)] Could not bind socket to 127.0.0.1:6004
+[0222/171151:ERROR:node_debugger.cc(86)] Cannot start debugger server
